requirements.txt
python-telegram-bot==20.7
Pillow
pytesseract
requests
schedule
supabase
python-dotenv
